Cleanup to multiple files, components
